% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bostonhouseprice}
\alias{bostonhouseprice}
\title{Boston House Price Data}
\format{
An object of class \code{data.frame} with 506 rows and 9 columns.
}
\usage{
bostonhouseprice
}
\description{
Boston House Price Data
}
\keyword{datasets}
