% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citycrime}
\alias{citycrime}
\title{City Crime Data}
\format{
A data frame with 50 rows and 7 variables:
\describe{
  \item{violent}{reported violent crime rate per 100,000 residents}
  \item{funding}{annual police funding per resident ($)}
  \item{hs}{percentage of people 25 years+ with 4 years of high school}
  \item{not_hs}{percentage of 16 to 19 year-olds not in highschool and not highschool graduates}
  \item{college}{percentage of 18 to 24 year-olds in college}
  \item{college4}{percentage of people 25 years+ with at least 4 years of college}
  \item{crime_rate}{total overall reported crime rate per 1 million residents}
  }
}
\source{
\url{https://hastie.su.domains/StatLearnSparsity_files/DATA/crime.txt}
}
\usage{
citycrime
}
\description{
Data relating to crime rates per one million residents in 50 U.S cities, taken
from Thomas (1990).
}
\references{
Thomas, G.S., 1990. The Rating Guide to Life in America's Small Cities.
Prometheus Books, 59 John Glenn Drive, Amherst, NY 14228-2197.

Hastie, T., Tibshirani, R. and Wainwright, M., 2015.
Statistical learning with sparsity: the lasso and generalizations. CRC press.
}
\keyword{datasets}
