% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Diabetes Data}
\format{
A data frame with 442 rows and 11 variables:
\describe{
  \item{AGE}{age of the patient}
  \item{SEX}{sex of the patient}
  \item{BMI}{body mass index of the patient}
  \item{BP}{blood pressure of the patient}
  \item{S1}{blood serum measurement 1}
  \item{S2}{blood serum measurement 2}
  \item{S3}{blood serum measurement 3}
  \item{S4}{blood serum measurement 4}
  \item{S5}{blood serum measurement 5}
  \item{S6}{blood serum measurement 6}
  \item{Y}{quantitative measure of disease progression one year after baseline}
  }
}
\source{
\url{https://CRAN.R-project.org/package=lars}
}
\usage{
diabetes
}
\description{
Data relating to a study of disease progression one year after baseline.
}
\references{
Efron, B., Hastie, T., Johnstone, I., Tibshirani, R., et al. (2004).
Least angle regression. The Annals of Statistics.
}
\keyword{datasets}
