% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothic_functions.R
\name{predict.smoothic}
\alias{predict.smoothic}
\title{Predict smoothic}
\usage{
\method{predict}{smoothic}(object, newdata, ...)
}
\arguments{
\item{object}{an object of class \dQuote{\code{smoothic}} which is the result
of a call to \code{\link{smoothic}}.}

\item{newdata}{new data object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a matrix containing the predicted values for the location mu and scale s
}
\description{
\code{predict} method class \dQuote{\code{smoothic}}
}
\examples{
# Sniffer Data --------------------
# MPR Model ----
results <- smoothic(
  formula = y ~ .,
  data = sniffer,
  family = "normal",
  model = "mpr"
)
predict(results)

}
\author{
Meadhbh O'Neill
}
