% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data_mpr}
\alias{sample_data_mpr}
\title{Sample Data for MPR Setting.}
\format{
A list with:
\describe{
  \item{x}{}
  \item{y}{}
  \item{beta_true}{}
  \item{alpha_true}{}
  ...
}
}
\usage{
sample_data_mpr
}
\description{
Sample Data for MPR Setting.
}
\keyword{datasets}
