% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothic_functions.R
\name{smoothic}
\alias{smoothic}
\title{Variable Selection Using a Smooth Information Criterion (SIC)}
\usage{
smoothic(
  x,
  y,
  model = "mpr",
  lambda = "log(n)",
  epsilon_1 = 10,
  epsilon_T = 1e-05,
  steps_T = 100,
  zero_tol = 1e-08,
  tol = 1e-08,
  max_it = 10000,
  initial_step = 10,
  max_step_it = 1000
)
}
\arguments{
\item{x}{Input matrix (unstandardized), of dimension nobs x nvars; each row
is an observation vector.}

\item{y}{Response variable.}

\item{model}{The type of regression to be implemented, either \code{model = "mpr"}
for multi-parameter regression, or \code{model = "spr"} for single parameter
regression (i.e., classical normal linear regression). Defaults to \code{model = "mpr"}.}

\item{lambda}{Value of penalty tuning parameter. Suggested values are
\code{"log(n)"} and \code{"2"} for the BIC and AIC respectively. Defaults to
\code{lambda ="log(n)"} for the BIC case.}

\item{epsilon_1}{Starting value for \eqn{\epsilon}-telescope. Defaults to 10.}

\item{epsilon_T}{Final value for \eqn{\epsilon}-telescope. Defaults to
\code{1e-05}.}

\item{steps_T}{Number of steps in \eqn{\epsilon}-telescope. Defaults to 100.}

\item{zero_tol}{Coefficients below this value are treated as being zero.
Defaults to \code{1e-08}.}

\item{tol}{Convergence tolerance for the optimization. Defaults to
\code{1e-08}.}

\item{max_it}{Maximum nuber of iterations to performed before the
optimization is terminated. Defaults to \code{1e+04}.}

\item{initial_step}{Initial step length for step halving in Newton-Raphson
algorithm. Defaults to 10.}

\item{max_step_it}{Maxiumum allowable number of steps to take for step
halving in Newton-Raphson algorithm. Defaults to \code{1e+03}.}
}
\value{

}
\description{
Implements the SIC \eqn{\epsilon}-telescope method,
returns coefficients and estimated standard errors (SEE). Note that the
function will scale the predictors to have unit variance, however, the final
estimates are converted back to their original scale.
}
\examples{
results_mpr <- smoothic_mpr(x = x, y = y, get_see = TRUE)
results_mpr$coefficients
results_mpr$see
}
