% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothic_functions.R
\name{smoothic_spr}
\alias{smoothic_spr}
\title{Single Parameter Smooth Information Criterion (SPR-SIC) Variable Selection
 Method}
\usage{
smoothic_spr(
  x,
  y,
  lambda = "log(n)",
  get_see = FALSE,
  epsilon_1 = 10,
  epsilon_T = 1e-05,
  steps_T = 100,
  zero_tol = 1e-08,
  theta_init = "lm",
  initial_step = 10,
  max_step_it = 1000,
  tol = 1e-08,
  max_it = 10000
)
}
\arguments{
\item{x}{Input matrix (unstandardized), of dimension nobs x nvars; each row
is an observation vector.}

\item{y}{Response variable.}

\item{get_see}{Logical flag for estimation of standard errors. Defaults to
\code{get_see=FALSE}.}

\item{epsilon_1}{Starting value for \eqn{\epsilon}-telescope. Defaults to 10.}

\item{epsilon_T}{Final value for \eqn{\epsilon}-telescope. Defaults to
\code{1e-05}.}

\item{steps_T}{Number of steps in \eqn{\epsilon}-telescope. Defaults to 100.}

\item{zero_tol}{Coefficients below this value are treated as being zero.
Defaults to \code{1e-08}.}

\item{theta_init}{Starting parameter values for the optimization. Defaults
to \code{theta_init="lm"}, which obtains the linear model estimates.
Alternatively, a vector of unscaled starting parameter values of length
(nvars + 2) can be used, i.e., initial parameter values for the location
parameter (including the intercept) and the variance term.}

\item{initial_step}{Initial step length for step halving in Newton-Raphson
algorithm. Defaults to 10.}

\item{max_step_it}{Maxiumum allowable number of steps to take for step
halving in Newton-Raphson algorithm. Defaults to \code{1e+03}.}

\item{tol}{Convergence tolerance for the optimization. Defaults to
\code{1e-08}.}

\item{max_it}{Maximum nuber of iterations to performed before the
optimization is terminated. Defaults to \code{1e+04}.}
}
\value{
A list with estimates and estimated standard errors.
\itemize{
  \item \code{coefficients} - vector of coefficients.
  \item \code{see} - vector of estimated standard errors
  (\code{NULL} if \code{get_see = FALSE})}
}
\description{
Implements the normal SPR-SIC \eqn{\epsilon}-telescope method,
returns coefficients and estimated standard errors (SEE). Note that the
function will scale the predictors to have unit variance, however, the final
estimates are converted back to their original scale.
}
\examples{
results_spr <- smoothic_spr(x = x, y = y, get_see = TRUE)
results_spr$coefficients
results_spr$see
}
