% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothic_functions.R
\name{summary.smoothic}
\alias{summary.smoothic}
\title{Summarising Smooth Information Criterion (SIC) Fits}
\usage{
\method{summary}{smoothic}(object)
}
\arguments{
\item{object}{an object of class \dQuote{\code{smoothic}} which is the result
of a call to \code{\link{smoothic}}.}
}
\value{
A list containing the following components:
\itemize{
  \item \code{model} the matched model from the \code{smoothic} object.
  \item \code{coefmat} a typical coefficient matrix whose columns are the
  estimated regression coefficients, estimated standard errors (SEE) and p-values.
  \item \code{plike} value of the penalized likelihood function.
  }
}
\description{
\code{summary} method class \dQuote{\code{smoothic}}
}
\examples{
# Prostate Cancer Data
x <- pcancer[, !(names_pcancer \%in\% c("lpsa"))]
y <- pcancer$lpsa

# MPR Model
results <- smoothic(x = x,
                    y = y,
                    model = "mpr")
summary(results)
}
\author{
Meadhbh O'Neill, \email{meadhbh.oneill@ul.ie}
}
