% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sniffer}
\alias{sniffer}
\title{Sniffer Data}
\format{
A data frame with 125 rows and 5 variables:
\describe{
  \item{tanktemp}{initial tank temperature (degrees F)}
  \item{gastemp}{temperature of the dispensed gasoline (degrees F)}
  \item{tankpres}{initial vapour pressure in the tank (psi)}
  \item{gaspres}{vapour pressure of the dispensed gasoline (psi)}
  \item{y}{hydrocarbons emitted (g)}
  }
}
\source{
\url{https://CRAN.R-project.org/package=alr4}
}
\usage{
sniffer
}
\description{
Data examining the factors that impact the amount of hydrocarbon vapour released
when gasoline is pumped into a tank.
}
\references{
Bedrick, E.J. (2000). Checking for lack of fit in linear models with parametric
variance functions. Technometrics 42 (3), 226–236.

Weisberg, S. (2014). Applied Linear Regression, 4th edition. Hoboken NJ: Wiley.
}
\keyword{datasets}
